+++ src/main/resources/plugins/kafka/prod2cons.py
@@ -1,6 +1,5 @@
 """
 Copyright (c) 2016 Cisco and/or its affiliates.
-
 This software is licensed to you under the terms of the Apache License, Version 2.0 (the "License").
 You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
 The code, technical concepts, and all information contained herein, are the property of
@@ -8,13 +7,10 @@ Cisco Technology, Inc. and/or its affiliated entities, under various laws includ
 international treaties, patent, and/or contract. Any use of the material herein must be in
 accordance with the terms of the License.
 All rights not expressly granted by the License are reserved.
-
 Unless required by applicable law or agreed to separately in writing, software distributed under
 the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND,
 either express or implied.
-
 Purpose:    Plugin for producer & consumer test to/from Kafka
-
 """
 
 import io
@@ -99,6 +95,11 @@ class Prod2Cons(object):
         writer = avro.io.DatumWriter(self.schema)
         for i in xrange(self.nbmsg):
             rawdata = "%s|%s" % (self.runtag, str(i))
+            quar_data = "%s|%s" % (self.runtag, str(i))
+            # Generating unique data for dataset health check
+            if i is 5:
+                rawdata = "%s|%s|%s" % (self.runtag, str(i), 'testavrodata')
+                quar_data = "%s|%s|%s" % (self.runtag, str(i), 'quardata')
             bytes_writer = io.BytesIO()
             encoder = avro.io.BinaryEncoder(bytes_writer)
             writer.write({"timestamp": TIMESTAMP_MILLIS(),
@@ -109,7 +110,8 @@ class Prod2Cons(object):
             raw_bytes = bytes_writer.getvalue()
             self.add_sent(i)
             self.producer.send_messages(self.topic, raw_bytes)
-            self.sent_msg += 1
+            self.producer.send_messages(self.topic, quar_data)
+            self.sent_msg += 2
         return 0
 
     def consumer_reset(self):
@@ -132,6 +134,11 @@ class Prod2Cons(object):
             readcount += 1
             try:
                 newmessage = message[1][3]
+                if newmessage.split('|')[0] == self.runtag:
+                    readvalid +=1
+                    self.add_rcv(int(newmessage.split('|')[1]))
+                    continue
+
                 bytes_reader = io.BytesIO(newmessage)
                 decoder = avro.io.BinaryDecoder(bytes_reader)
                 reader = avro.io.DatumReader(self.schema)
@@ -153,5 +160,4 @@ class Prod2Cons(object):
         if readcount == self.nbmsg and readvalid == self.nbmsg:
             LOGGER.debug("consumer : test run ok")
             avg_ms = self.average_ms()
-
         return TestbotResult(self.sent_msg, readvalid, readnotvalid, avg_ms)